requirements.txt
openai
beautifulsoup4
python-dotenv
lxml
